Use SVG for icons instead of PNG
As an alternative to #270 based on the suggestion in [this comment](https://github.com/maximebf/php-debugbar/pull/270#issuecomment-215638893) I’ve provided a CSS file with embedded SVGs rather than the included font awesome.

There are pros and cons to each approach, but I’d love to see vector icons in the default debug bar.