Fix incorrect duplicate statement counter in PDO collector (#323)
The PDO collector assumes that the number of duplicate statements is
equal to the number of total statements minus the number of unique
statement query strings that had more than one query.  This leads to
non-sensical statements like the following:

    5 statements were executed, 4 of which were duplicated, 1 unique

when the following statements were executed.  (Assume each letter
represents a particular query)

* A
* A
* B
* C
* D

Clearly there are not 4 duplicate statements under any reasonable
interpretation.  After this code change, it would more correctly say for
the above example:

    5 statements were executed, 2 of which were duplicates, 3 unique